Fixing broken test classes
