+++ src/test/java/org/imsglobal/caliper/lti/ResourceLinkClaim.java
@@ -19,7 +19,7 @@
 package org.imsglobal.caliper.lti;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.imsglobal.caliper.v1p1.events.ViewEventViewedFedSessionTest;
+import org.imsglobal.caliper.v1p1.events.ViewEventViewedDocumentFedSessionTest;
 
 public class ResourceLinkClaim {
     @JsonProperty("id")
